% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_profile.R
\name{Barplot_Pathway2Species}
\alias{Barplot_Pathway2Species}
\title{Barplot for humann2 profile}
\usage{
Barplot_Pathway2Species(dat, group, pathwayID)
}
\arguments{
\item{dat}{full combined profile}

\item{group}{group information with one column}

\item{pathwayID}{pathway ID}
}
\value{
value      vector of standarized species name
}
\description{
A function to implement barplot in humann2
}
