% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Centering}
\usage{
center(dataSet, names = c("gender", "treatment", "replicate"))
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{names}{A vector of strings (default = c("gender", "treatment", "replicate"))
specifying the names of the attribute columns.}
}
\value{
The centered data.
}
\description{
Apply centering to the data.
}
\details{
The function executes the following:
\enumerate{
\item Plots the mean-variance relationship.
\item Centers the data.
\item Plots the mean-variance relationship again for comparison.
}
}
