% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_DataFram.R
\name{CombineTidyTables}
\alias{CombineTidyTables}
\title{Combine Dataframes}
\usage{
CombineTidyTables(datlst, by, method = dplyr::full_join, use_list_name = TRUE)
}
\arguments{
\item{by}{which column used to combine}

\item{method}{combine method in dplyr packages used to combine dataframes. Default is full_join. Supported methods including full_join,left_join,right_join,inner_jion}

\item{use_list_name}{if use list name replace original names.}

\item{datalst}{input list of dataframes}
}
\value{
dataframe
}
\description{
Inplementation of combine tables function in plyr
}
