% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_profile.R
\name{FilterMetaphlan2Profile}
\alias{FilterMetaphlan2Profile}
\title{Filter metaphlan2 profile}
\usage{
FilterMetaphlan2Profile(
  pro,
  include = c("k__Bacteria", "k__Archaea"),
  remove_unclassified = TRUE,
  remove_noname = TRUE,
  normalize = TRUE
)
}
\arguments{
\item{pro}{input splited metaphlan2 profile in a rank, e.g. genus}

\item{include}{any kingdom included, default k__Bacteria and K__Archaea. The prefix "K__" is needed.}

\item{remove_unclassified}{if remove unclasssified features in profiles.}

\item{remove_noname}{if remove features with noname.}

\item{normalize}{if normalization after above process.}
}
\value{
dataframe
}
\description{
Filter metaphlan2 profile
}
