% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filterNA}
\alias{filterNA}
\title{Filtering NA's post-imputation}
\usage{
filterNA(dataSet, saveRm = TRUE)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{saveRm}{A boolean (default = TRUE) specifying whether to save removed data to
current working directory.}
}
\value{
The filtered data.
}
\description{
Remove compounds with NA values.
}
\details{
If compounds that do not meet the imputation requirement are removed, a .csv file is
created with the removed data.
}
