% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filterOutIn}
\alias{filterOutIn}
\title{Filtering compounds}
\usage{
filterOutIn(
  dataSet,
  listName = c(),
  regexName = c(),
  removeList = TRUE,
  saveRm = TRUE
)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{listName}{A character vector of compounds to select or remove.}

\item{regexName}{A character vector specifying compounds for regular expression pattern
matching to select or remove.}

\item{removeList}{A boolean (default = TRUE) specifying whether the list of compounds
should be removed or selected.
\itemize{
\item TRUE: Remove the list of compounds from the data.
\item FALSE: Remove all compounds not in the list from the data.
}}

\item{saveRm}{A boolean (default = TRUE) specifying whether to save removed data to
current working directory. This option only works when \code{removeList = TRUE}.}
}
\value{
The filtered data.
}
\description{
Apply a series of filtering steps to the data set.
}
\details{
If both \code{listName} and \code{regexName} are provided, the compound names selected
or removed will be the union of those specified in \code{listName} and those matching
the regex pattern in \code{regexName}.
}
