% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute.knn_trunc}
\alias{impute.knn_trunc}
\title{Imputation by the truncated k-nearest neighbors algorithm}
\usage{
impute.knn_trunc(
  dataSet,
  names = c("gender", "treatment", "replicate"),
  k = 10
)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{names}{A vector of strings (default = c("gender", "treatment", "replicate"))
specifying the names of the attribute columns.}

\item{k}{An integer (default = 10) indicating the number of neighbors to be used in the
imputation.}
}
\value{
The imputed data.
}
\description{
Apply imputation to the data by the truncated k-nearest neighbors algorithm
\insertCite{shah2017distribution}{metastat}.
}
\references{
\insertAllCited{}
}
