% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute.min_global}
\alias{impute.min_global}
\title{Imputation by the global minimum}
\usage{
impute.min_global(dataSet, names = c("gender", "treatment", "replicate"))
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{names}{A vector of strings (default = c("gender", "treatment", "replicate"))
specifying the names of the attribute columns.}
}
\value{
The imputed data.
}
\description{
Apply imputation to the data by the minimum measured value from any compound found
within the entire data.
}
