% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute.min_local}
\alias{impute.min_local}
\title{Imputation by the local minimum}
\usage{
impute.min_local(dataSet, reqPercentPresent = 0.51)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{reqPercentPresent}{A scalar (default = 0.51) specifying the required percent of
values that must be present in a given compound by condition combination for values to
be imputed.}
}
\value{
The imputed data.
}
\description{
Apply imputation to the data by the minimum measured value for that compound in that
condition.
}
