% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute.pca_prob}
\alias{impute.pca_prob}
\title{Imputation by probabilistic principal components analysis}
\usage{
impute.pca_prob(
  dataSet,
  names = c("gender", "treatment", "replicate"),
  nPcs = NULL,
  maxIterations = 1000,
  seed = 362436069
)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{names}{A vector of strings (default = c("gender", "treatment", "replicate"))
specifying the names of the attribute columns.}

\item{nPcs}{An integer specifying the number of principal components to calculate. The
default is set to the minimum between the number of samples and the number of proteins.}

\item{maxIterations}{An integer (default = 1000) specifying the maximum number of
allowed iterations.}

\item{seed}{An integer (default = 362436069) specifying the seed used for the random
number generator for reproducibility.}
}
\value{
The imputed data.
}
\description{
Apply imputation to the data by probabilistic principal components analysis
\insertCite{stacklies2007pcamethods}{metastat}.
}
\references{
\insertAllCited{}
}
