% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{INT}
\alias{INT}
\title{INT transform
a function to permform INT-transformation}
\usage{
INT(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{

}
\description{
INT transform
a function to permform INT-transformation
}
