% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnimputate.R
\name{knnimpute}
\alias{knnimpute}
\title{kNN Missing Value Imputation}
\usage{
knnimpute(x, k = 3)
}
\arguments{
\item{x}{An array-like object that contains the data with NaNs.}

\item{k}{A positive integer excluding 0 (default: 3). The number of nearest neighbors to use.}
}
\value{
An array-like object corresponding to x with NaNs imputed.
}
\description{
Impute missing values in a matrix using kNN (k-Nearest Neighbors) imputation based on Euclidean distance.
}
\examples{
x <- matrix(c(1, 2, 3, NA, 5, 6, 7, 8, 9), ncol = 3)
imputed_data <- knnimpute(x, k = 2)

}
