% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparionStat.R
\name{KW_phenotypes}
\alias{KW_phenotypes}
\title{KW test}
\usage{
KW_phenotypes(dat, grp)
}
\arguments{
\item{dat}{, input data}

\item{grp}{, input groups with one column}
}
\value{
dataframe
}
\description{
This function perform Kruskal-Walis test among multiple gorups
}
