% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{ma}
\alias{ma}
\title{The input for MA plot}
\usage{
ma(dataSet, condCol = "merged_condition", cond)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{condCol}{A string (default = "merged_condition") specifying the column name in
\code{dataSet} that contains the conditions to compared.}

\item{cond}{A string specifying which two conditions to compare. The order is
important, as the second condition serves as the reference for comparison. When there
are two conditions in \code{condCol} of \code{dataSet} and this argument is not
specified, the \code{cond} will automatically be selected by sorting the unique
values alphabetically and in ascending order.}
}
\value{
A data frame containing the compound-wise averages within each condition.
}
\description{
Calculate the input for generating an MA plot.
}
\details{
The second condition serves as the reference for comparison. The rows are ordered by
\code{cond}. Specifically, the first row corresponds to the compound-wise average of
the first condition, and the second row corresponds to the second condition.
}
