% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{mediation_analysis}
\alias{mediation_analysis}
\title{Customized mediation analysis}
\usage{
mediation_analysis(data, x, y, m, adj)
}
\arguments{
\item{data}{input data}

\item{x}{x}

\item{y}{y}

\item{m}{mediation variable}

\item{adj}{adjust variable}
}
\value{
dataframe
}
\description{
Customized mediation analysis.
fit1: m ~ x + adj
fit2: y ~ x + m + adj
}
