% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{mod_t_test}
\alias{mod_t_test}
\title{Moderated t-test}
\usage{
mod_t_test(dataSet, condCol = "merged_condition", cond)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{condCol}{A string (default = "merged_condition") specifying the column name in \code{dataSet} that contains
the conditions to compared.}

\item{cond}{A string specifying which two conditions to compare. The order is
important, as the second condition serves as the reference for comparison. When there
are two conditions in \code{condCol} of \code{dataSet} and this argument is not
specified, the \code{cond} will automatically be selected by sorting the unique
values alphabetically and in ascending order.}
}
\value{
A data frame containing the differences in means and p-values for each
compound between the two conditions.
}
\description{
Moderated t-test
}
