% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_heatmap.R
\name{mypheatmap}
\alias{mypheatmap}
\title{customize pheatmap}
\usage{
mypheatmap(scc, ReAdjPvalLowerTri = FALSE, FilterRow = F, FilterCol = F, ...)
}
\arguments{
\item{scc}{output of scc_self, scc_between, pcc_self, pcc_between}

\item{ReAdjPvalLowerTri}{if re-adjust pvalue (lower tri)}

\item{FilterRow}{FilterRow}

\item{FilterCol}{FilterCol}

\item{...}{other parameters passed to pheatmap}
}
\value{

}
\description{
customize pheatmap
}
