% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalization}
\usage{
normalize(dataSet, method = "quantile")
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{method}{A string (default = "quant") specifying the method of normalization to
apply:
\enumerate{
\item Row-wise (sample-wise) normalization:
\itemize{
\item "mean": Samples are normalized to the mean value of all variables for a given
sample.
\item "median": samples are normalized to the median value of all variables for a given
sample.
\item "quantile": Quantile normalization.
}
\item Column-wise (metabolite-wise) normalization:
\itemize{
\item "meanCtr": Mean centering.
\item "medianCtr": Median centering.
\item "auto": Auto scaling.
\item "level": Level scaling.
\item "pareto": Pareto scaling.
\item "range": Range scaling.
\item "vast": Vast scaling.
}
}}
}
\value{
The normalized data.
}
\description{
Apply a specified type of normalization to the data.
}
\details{
Quantile normalization is generally recommended. Mean and median normalization are
going to be included as popular previous methods. No normalization is not recommended.
Boxplots are also generated for before and after the normalization to give a visual
indicator of the changes.
}
