% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{PCA Score and Loadings Plot}
\arguments{
\item{X}{An array-like object that contains the input data.}

\item{pcx}{A positive integer representing the principal component for the x-axis (default: 1).}

\item{pcy}{A positive integer representing the principal component for the y-axis (default: 2).}

\item{group_label}{A vector specifying group labels for coloring points in the score plot (default: NULL).}

\item{sample_label}{A vector specifying labels for individual points in the score plot (default: NULL).}

\item{peak_label}{A vector specifying labels for individual points in the loadings plot (default: NULL).}

\item{plot_ci}{A logical indicating whether to plot 95% confidence ellipses in the score plot (default: TRUE).}

\item{grid_line}{A logical indicating whether to show grid lines in both plots (default: TRUE).}
}
\description{
Creates a PCA scores and loadings plot using ggplot2 and FactoMineR.
}
\examples{
X <- matrix(rnorm(100), ncol = 10)
pca_plot(X, pcx = 1, pcy = 2, group_label = rep(1:5, each = 20), sample_label = rep(1:10, each = 10), plot_ci = TRUE, grid_line = TRUE)

}
