% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{pcc_between}
\alias{pcc_between}
\title{Partial Spearman's correlation}
\usage{
pcc_between(dat1, dat2, dat_adj)
}
\arguments{
\item{dat1}{input data1}

\item{dat2}{input data2}

\item{dat_adj}{input adjustment data}
}
\value{
dataframe
}
\description{
This function perform partial spearman's correlation between two dataframe
}
