% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{pcc_self}
\alias{pcc_self}
\title{Partial Spearman's correlation}
\usage{
pcc_self(dat, dat_adj)
}
\arguments{
\item{dat}{input data}

\item{data_adj}{input adjustment}
}
\value{
dataframe
}
\description{
This function perform partial spearman's correlation in one dataframe
}
