% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Loading and reformatting of metabolomics data}
\usage{
preprocess(
  fileName,
  dataSet = NULL,
  names = c("gender", "treatment", "replicate")
)
}
\arguments{
\item{fileName}{The name of the .csv file containing metabolomics data (including the
path to the file, if needed).}

\item{dataSet}{The raw data set, if already loaded in R.}

\item{names}{A vector of strings (default = c("gender", "treatment", "replicate"))
specifying the names of the attribute columns.}
}
\value{
A 2d dataframe.
}
\description{
Read a data file, select columns necessary for analysis, and return the reformatted data.
}
\details{
The function executes the following:
\enumerate{
\item Reads the file.
\item Provides summary statistics and a histogram of all values reported in the data set.
\item Re-formats the data to present individual compounds as columns.
\item Stores the data as a \code{data.frame} and prints the levels of attributes to the
user.
}
}
