% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_Maaslin2.R
\name{ReAdjustBHMaaslin2}
\alias{ReAdjustBHMaaslin2}
\title{read maaslin2 results and re-adjusted(BH)}
\usage{
ReAdjustBHMaaslin2(Path)
}
\arguments{
\item{path}{specify a directory of the maaslin results}
}
\value{
dataframe
}
\description{
read maaslin2 results and re-adjusted(BH) in a specified directory
}
