% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale}
\alias{scale}
\title{Centering and/or scaling}
\usage{
scale(
  dataSet,
  names = c("gender", "treatment", "replicate"),
  center = TRUE,
  scale = "auto",
  mu = NULL,
  sigma = NULL
)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{names}{A vector of strings (default = c("gender", "treatment", "replicate"))
specifying the names of the attribute columns.}

\item{center}{A boolean (default = TRUE) specifying whether to apply centering.}

\item{scale}{A string (default = "auto") specifying which method of scaling to apply:
\enumerate{
\item "none" No scaling will be applied.
\item "auto": Auto scaling.
\item "level": Level scaling.
\item "pareto": Pareto scaling.
\item "range": Range scaling.
\item "vast": Vast scaling.
}}

\item{mu}{A numeric vector specifying the centering reference. The default centering
reference is set to the mean value.}

\item{sigma}{A numeric vector specifying the scaling reference based on the corresponding
definition. The default scaling reference is set to the standard deviation when
\code{center = TRUE} and the adjusted root mean square when \code{center = FALSE}.}
}
\value{
The centered and/or scaled data.
}
\description{
Apply centering and/or scaling to the data.
}
\details{
The function executes the following:
\enumerate{
\item Plots the mean-variance relationship.
\item Centers and/or scales the data.
\item Plots the mean-variance relationship again for comparison.
}
}
