% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{scc_between}
\alias{scc_between}
\title{Spearman's correlation}
\usage{
scc_between(dat1, dat2)
}
\arguments{
\item{dat1}{input data1}

\item{dat2}{input data2}
}
\value{
dataframe
}
\description{
This function perform spearman's correlation between two dataframe
}
