% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{scc_self}
\alias{scc_self}
\title{Spearman's correlation}
\usage{
scc_self(dat)
}
\arguments{
\item{dat}{input data}
}
\value{
dataframe
}
\description{
This function perform spearman's correlation in one dataframe
}
