% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_profile.R
\name{SplitMetaphlan2Profile}
\alias{SplitMetaphlan2Profile}
\title{Split metaphlan2 profile by given rank}
\usage{
SplitMetaphlan2Profile(pro, rank = "a")
}
\arguments{
\item{pro}{input combined metaphlan2 profile}

\item{rank}{input a given rank from [a,p,c,o,f,g,s],default a for all ranks}
}
\value{
dataframe or list
}
\description{
Split metaphlan2 profile by given rank
}
