% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_profile.R
\name{standarizeSpeciesName}
\alias{standarizeSpeciesName}
\title{Standarize species name in  metaphlan2 profile}
\usage{
standarizeSpeciesName(x)
}
\arguments{
\item{x}{vector of species name}

\item{italic}{if use italic face}
}
\value{
value      vector of standarized species name
}
\description{
A function to standarize species names in metaphlan2 profile
}
