% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparionStat.R
\name{wilcox.customized}
\alias{wilcox.customized}
\title{customized wilcoxon test}
\usage{
wilcox.customized(dat, grp, type = "phenotype")
}
\arguments{
\item{dat}{, input data}

\item{grp}{, input groups with one column}
}
\value{
dataframe
}
\description{
This function perform wilcox test between gorups
}
