% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparionStat.R
\name{wilcox.Paired}
\alias{wilcox.Paired}
\title{customized paired wilcoxon-rank sum test}
\usage{
wilcox.Paired(dat, grp, paired, group)
}
\arguments{
\item{dat}{input data}

\item{grp}{groups with two column, "paired" for patient ID and "group" for treatment groups}

\item{paired}{which column used as patiend ID}

\item{group}{which column used as treatment groups}
}
\value{
dataframe
}
\description{
The function to perform paired wilcoxon-rank sum test between gorups
}
