% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute.mice_cart}
\alias{impute.mice_cart}
\title{Imputation by classification and regression trees}
\usage{
impute.mice_cart(
  dataSet,
  names = c("gender", "treatment", "replicate"),
  m = 5,
  seed = 362436069
)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{names}{A vector of strings (default = c("gender", "treatment", "replicate"))
specifying the names of the attribute columns.}

\item{m}{An integer (default = 5) specifying the number of multiple imputations.}

\item{seed}{An integer (default = 362436069) specifying the seed used for the random
number generator for reproducibility.}
}
\value{
The imputed data.
}
\description{
Apply imputation to the data by classification and regression trees
\insertCite{breiman1984classification,doove2014recursive,van2018flexible}{metastat}.
}
\references{
\insertAllCited{}
}
