% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_heatmap.R
\name{mypheatmap}
\alias{mypheatmap}
\title{customize pheatmap}
\usage{
mypheatmap(
  scc,
  ReAdjPvalLowerTri = FALSE,
  FilterRow = F,
  FilterCol = F,
  cellheight = 15,
  cellwidth = 20,
  fontsize = 10,
  fontsize_number = 10,
  cluster_rows = T,
  cluster_cols = T,
  ...
)
}
\arguments{
\item{scc}{output of scc_self, scc_between, pcc_self, pcc_between}

\item{ReAdjPvalLowerTri}{if re-adjust pvalue (lower tri)}

\item{FilterRow}{FilterRow}

\item{FilterCol}{FilterCol}

\item{cellheight}{cellheight}

\item{cellwidth}{cellwidth}

\item{fontsize}{fontsize}

\item{fontsize_number}{fontsize_number}

\item{cluster_rows}{cluster_rows}

\item{cluster_cols}{cluster_cols}

\item{...}{other parameters passed to pheatmap}
}
\value{

}
\description{
customize pheatmap
}
