% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalization}
\usage{
normalize(
  dataSet,
  names = c("gender", "treatment", "replicate"),
  method = "quant"
)
}
\arguments{
\item{dataSet}{A data frame containing the data signals.}

\item{names}{A vector of strings (default = c("gender", "treatment", "replicate"))
specifying the names of the attribute columns.}

\item{method}{A string (default = "quant") specifying the method of normalization to
apply:
\enumerate{
\item Centering:
\itemize{
\item "mean": Mean centering.
\item "median": Median centering.
}
\item Scaling:
\itemize{
\item "auto": Auto scaling.
\item "level": Level scaling.
\item "pareto": Pareto scaling.
\item "range": Range scaling.
\item "vast": Vast scaling.
}
\item "quant": Quantile normalization.
\item "none": None.
}}
}
\value{
A normalized 2d dataframe.
}
\description{
Apply a specified type of normalization to the data.
}
\details{
Quantile normalization is generally recommended. Mean and median normalization are
going to be included as popular previous methods. No normalization is not recommended.
Boxplots are also generated for before and after the normalization to give a visual
indicator of the changes.
}
