% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{PCbiplot}
\alias{PCbiplot}
\title{Title}
\usage{
PCbiplot(
  PC,
  x = "PC1",
  y = "PC2",
  Group = NULL,
  main = NULL,
  point_size = 3,
  point_alpha = 1
)
}
\arguments{
\item{PC}{output of prcomp function}

\item{x}{PC1 to display}

\item{y}{PC2 to display}

\item{Group}{dataframe with one column to group}

\item{main}{main title}

\item{point_size}{size of point}

\item{point_alpha}{alpha of point}
}
\value{

}
\description{
Title
}
