% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{PCbiplot}
\alias{PCbiplot}
\title{Title}
\usage{
PCbiplot(PC, x = "PC1", y = "PC2", Group)
}
\arguments{
\item{PC}{output of prcomp function}

\item{x}{PC1 to display}

\item{y}{PC2 to display}

\item{Group}{dataframe with one column to group}
}
\value{

}
\description{
Title
}
