% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{PermanovaMulti}
\alias{PermanovaMulti}
\title{Title}
\usage{
PermanovaMulti(x, distance, ...)
}
\arguments{
\item{x}{datafram input}

\item{distance}{distance or matrix}

\item{...}{Other parameters passsed to adonis}
}
\value{

}
\description{
Title
}
