% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale}
\alias{scale}
\title{Scale a matrix or data frame}
\usage{
scale(
  x,
  axis = 1,
  ddof = 1,
  method = "auto",
  mu = "default",
  sigma = "default",
  return_mu_sigma = FALSE
)
}
\arguments{
\item{x}{An array-like object that contains the data.}

\item{axis}{An integer or \code{NULL} (default \code{1}). The axis along which to operate.}

\item{ddof}{An integer (default \code{1}). The degrees of freedom correction.}

\item{method}{A string (default \code{"auto"}). Method used to scale \code{x}.
Accepted methods are \code{"auto"}, \code{"pareto"}, \code{"vast"}, \code{"level"}, or \code{"range"}.}

\item{mu}{A number or \code{"default"} (default \code{"default"}).
If \code{mu} is provided, it is used; otherwise, it is calculated.}

\item{sigma}{A number or \code{"default"} (default \code{"default"}).
If \code{sigma} is provided, it is used; otherwise, it is calculated.}

\item{return_mu_sigma}{A logical (default \code{FALSE}).
If \code{TRUE}, \code{mu} and \code{sigma} are returned instead of \code{z}.}
}
\value{
If \code{return_mu_sigma = FALSE}, an array-like object containing the scaled data.
  If \code{return_mu_sigma = TRUE}, a list with \code{mu} and \code{sigma}.
}
\description{
Scales the input matrix or data frame using various scaling methods.
}
\examples{
x <- matrix(rnorm(20), ncol = 2)
scaled_data <- scale(x)

}
