% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Title}
\usage{
scatterplot(dat, x, y)
}
\arguments{
\item{dat}{dataframe, input}

\item{x}{string, lab for x}

\item{y}{string, lab for y}
}
\value{

}
\description{
Title
}
