% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{scatterplot
customized scatterplot}
\usage{
scatterplot(
  dat,
  x,
  y,
  group = NULL,
  PCC = F,
  adj = NULL,
  size.points = 1,
  alpha.points = 0.5,
  size.smooth = 1,
  size.lab = 3
)
}
\arguments{
\item{dat}{dataframe, input}

\item{x}{string, lab for x}

\item{y}{string, lab for y}

\item{group}{string, lab for grouping, default NULL}

\item{PCC}{if permform ppcor::pcor.test}

\item{adj}{string, lab for confounder in pcor.test}
}
\value{

}
\description{
scatterplot
customized scatterplot
}
